pyproject.toml
[project]
name = "binance-square-engine"
version = "0.1.0"
description = "AI-powered Binance Square growth engine: signal detection → viral content generation → visual cards → auto-publishing → learning loop"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "KinzoTech" }]

dependencies = [
    # HTTP + async
    "httpx[http2]>=0.27.0",
    "tenacity>=8.5.0",
    # Browser automation (image upload, fallback publisher)
    "playwright>=1.45.0",
    # LLM
    "google-generativeai>=0.7.0",
    # Scheduler
    "apscheduler>=3.10.4",
    # Database
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.2",
    # Config / models
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    # CLI
    "typer>=0.12.3",
    "rich>=13.7.1",
    # Logging
    "structlog>=24.2.0",
    # Charts (for chart card rendering data)
    "pandas>=2.2.2",
    # FastAPI dashboard (optional but recommended for monitoring)
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "black>=24.4.2",
    "mypy>=1.10.0",
]

[project.scripts]
bse = "engine.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "C4"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
